Title: 정사각 알파벳 출력하기, Time: 134ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241021/정사각 알파벳 출력하기/print-alpabet-in-square.java
@@ -8,7 +8,7 @@ public class Main {
         for(int i=1; i<=n; i++){
             for(int j=1; j<=n; j++){
                 System.out.print(a);
-                (int)a++;
+                a++;
             }
             System.out.println();
         }
